+++ CHANGELOG.md
@@ -22,6 +22,12 @@ All notable changes to this project will be documented here. For more details, v
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.0.1] - 2026-03-17
+
+### Fixed
+
+- Update website salt (Gofile)
+
 ## [9.0.0] - 2026-03-09
 
 ⚠️**BREAKING CHANGES**
@@ -54,6 +60,7 @@ Several config options:
 - Download of multipage profile albums (Chevereto)
 - Handle `redirect/to` links (Xenforo)
 - Download of files with broken thumbnails (bunkr)
+- Send custom one-time-use headers (GoFile)
 - Download subfolder from filesystem URLs (pixeldrain)
 
 ## [8.10.0] - 2026-01-26